add signInView and registrationView
