+++ LoveCalendar/LoveCalendar/Flows/Registration/Screen/RegistrationViewController.swift
@@ -9,75 +9,7 @@ import UIKit
 import Combine
 
 class RegistrationViewController: UIViewController, FlowController {
-    private lazy var regLabel: UILabel = {
-        let label = UILabel()
-        label.text = "Регистрация"
-        label.font = .systemFont(ofSize: 20, weight: .bold)
-        label.textColor = UIColor(named: "LabelTextColor")
-        return label
-    }()
-
-    private lazy var emailTextField: UITextField = {
-        let textField = UITextField()
-        textField.placeholder = "Почта"
-        textField.borderStyle = .roundedRect
-        textField.keyboardType = .emailAddress
-        textField.returnKeyType = .next
-        textField.delegate = self
-        textField.textColor = UIColor(named: "LabelTextColor")
-        textField.font = .systemFont(ofSize: 18)
-        return textField
-    }()
-
-    private lazy var passwordTextField: UITextField = {
-        let textField = UITextField()
-        textField.placeholder = "Пароль"
-        textField.borderStyle = .roundedRect
-        textField.isSecureTextEntry = true
-        textField.returnKeyType = .next
-        textField.delegate = self
-        textField.textColor = UIColor(named: "LabelTextColor")
-        textField.font = .systemFont(ofSize: 18)
-        return textField
-    }()
-
-    private lazy var passwordConfirmationTextField: UITextField = {
-        let textField = UITextField()
-        textField.placeholder = "Подтвердите пароль"
-        textField.borderStyle = .roundedRect
-        textField.isSecureTextEntry = true
-        textField.returnKeyType = .done
-        textField.delegate = self
-        textField.textColor = UIColor(named: "LabelTextColor")
-        textField.font = .systemFont(ofSize: 18)
-        return textField
-    }()
-
-    private lazy var errorsLabel: UILabel = {
-        let label = UILabel()
-        label.font = .systemFont(ofSize: 14)
-        label.textColor = .red
-        label.numberOfLines = 0
-        return label
-    }()
-
-    private lazy var regButton: UIButton = {
-        let button = UIButton(type: .roundedRect)
-        button.layer.cornerRadius = 8
-        button.titleLabel?.font = .systemFont(ofSize: 18)
-        button.setTitle("Зарегестрироваться", for: .normal)
-        button.setTitleColor(.white, for: .normal)
-        button.backgroundColor = UIColor(named: "ButtonTextColor")
-
-        let action = UIAction { [weak self] _ in
-            guard let self else { return }
-            self.viewModel.signUpUser(emailTextField.text, passwordTextField.text, passwordConfirmationTextField.text)
-        }
-        button.addAction(action, for: .touchUpInside)
-
-        return button
-    }()
-
+    private let registrationView = RegistrationView(frame: .zero)
     var completionHandler: (() -> Void)?
     private let viewModel: RegistrationViewModel
     var cancellables = Set<AnyCancellable>()
@@ -91,48 +23,21 @@ class RegistrationViewController: UIViewController, FlowController {
         fatalError("init(coder:) has not been implemented")
     }
 
+    override func loadView() {
+        view = registrationView
+    }
+
     override func viewDidLoad() {
         super.viewDidLoad()
         view.backgroundColor = .white
-        setLayout()
+        registrationView.delegate = self
         setBindings()
     }
 }
 
-extension RegistrationViewController {
-    private func setLayout() {
-        let stackView = UIStackView(arrangedSubviews: [
-            emailTextField,
-            passwordTextField,
-            passwordConfirmationTextField
-        ])
-        stackView.axis = .vertical
-        stackView.spacing = 20
-
-        addSubviews(regLabel, stackView, errorsLabel, regButton)
-
-        regLabel.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.top.equalToSuperview().offset(view.frame.height * 0.2)
-        }
-
-        stackView.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.width.equalTo(300)
-            make.top.equalTo(regLabel).offset(view.frame.height * 0.1)
-        }
-
-        errorsLabel.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.top.equalTo(stackView.snp_bottomMargin).offset(50)
-            make.left.right.equalTo(stackView)
-        }
-
-        regButton.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.width.equalTo(200)
-            make.top.equalTo(errorsLabel).offset(view.frame.height * 0.1)
-        }
+extension RegistrationViewController: RegistrationViewDelegate {
+    func didPressSignUpButton(_ email: String?, _ password: String?, _ passwordConfirmation: String?) {
+        self.viewModel.signUpUser(email, password, passwordConfirmation)
     }
 }
 
@@ -151,24 +56,8 @@ extension RegistrationViewController {
             .receive(on: DispatchQueue.main)
             .sink { [weak self] errorStringFormatted in
                 guard let self = self else { return }
-                self.errorsLabel.text = errorStringFormatted
+                self.registrationView.setErrors(errorStringFormatted)
             }
             .store(in: &cancellables)
-    }
-}
-
-extension RegistrationViewController: UITextFieldDelegate {
-    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
-        switch textField {
-        case emailTextField:
-            passwordTextField.becomeFirstResponder()
-        case passwordTextField:
-            passwordConfirmationTextField.becomeFirstResponder()
-        case passwordConfirmationTextField:
-            view.endEditing(true)
-        default:
-            break
-        }
-        return true
     }
 }
