+++ LoveCalendar/LoveCalendar/Flows/SignIn/Screen/SignInView.swift
@@ -1,0 +1,140 @@
+//
+//  SignInView.swift
+//  LoveCalendar
+//
+//  Created by kerik on 14.04.2024.
+//
+
+import UIKit
+
+protocol SignViewDelegate: AnyObject {
+    func didPressSignInButton(_ email: String?, _ password: String?)
+}
+
+class SignInView: UIView {
+    weak var delegate: SignViewDelegate?
+
+    private lazy var signInLabel: UILabel = {
+        let label = UILabel()
+        label.text = "Вход"
+        label.font = .systemFont(ofSize: 20, weight: .bold)
+        label.textColor = UIColor(named: "LabelTextColor")
+        return label
+    }()
+
+    private lazy var emailTextField: UITextField = {
+        let textField = UITextField()
+        textField.placeholder = "Почта"
+        textField.borderStyle = .roundedRect
+        textField.keyboardType = .emailAddress
+        textField.returnKeyType = .next
+        textField.delegate = self
+        textField.textColor = UIColor(named: "LabelTextColor")
+        textField.font = .systemFont(ofSize: 18)
+        return textField
+    }()
+
+    private lazy var passwordTextField: UITextField = {
+        let textField = UITextField()
+        textField.placeholder = "Пароль"
+        textField.borderStyle = .roundedRect
+        textField.isSecureTextEntry = true
+        textField.returnKeyType = .next
+        textField.delegate = self
+        textField.textColor = UIColor(named: "LabelTextColor")
+        textField.font = .systemFont(ofSize: 18)
+        return textField
+    }()
+
+    private lazy var errorsLabel: UILabel = {
+        let label = UILabel()
+        label.font = .systemFont(ofSize: 14)
+        label.textColor = .red
+        label.numberOfLines = 0
+        return label
+    }()
+
+    private lazy var signInButton: UIButton = {
+        let button = UIButton(type: .roundedRect)
+        button.layer.cornerRadius = 20
+        button.titleLabel?.font = .systemFont(ofSize: 18, weight: .bold)
+        button.setTitle("Войти", for: .normal)
+        button.setTitleColor(.white, for: .normal)
+        button.backgroundColor = UIColor(named: "ButtonTextColor")
+
+        let action = UIAction { [weak self] _ in
+            guard let self else { return }
+            self.delegate?.didPressSignInButton(emailTextField.text, passwordTextField.text)
+        }
+        button.addAction(action, for: .touchUpInside)
+
+        return button
+    }()
+
+    private lazy var stackView: UIStackView = createStackView()
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        addSubviews(signInLabel, stackView, errorsLabel, signInButton)
+        setLayout()
+    }
+
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
+
+extension SignInView {
+    func setErrors(_ errors: String) {
+        errorsLabel.text = errors
+    }
+}
+
+extension SignInView {
+    private func createStackView() -> UIStackView {
+        let stackView = UIStackView(arrangedSubviews: [emailTextField, passwordTextField])
+        stackView.axis = .vertical
+        stackView.spacing = 20
+
+        return stackView
+    }
+    private func setLayout() {
+        signInLabel.snp.makeConstraints { make in
+            make.centerX.equalToSuperview()
+            make.top.equalToSuperview().offset(150)
+        }
+
+        stackView.snp.makeConstraints { make in
+            make.centerX.equalToSuperview()
+            make.width.equalTo(300)
+            make.top.equalTo(signInLabel).offset(100)
+        }
+
+        errorsLabel.snp.makeConstraints { make in
+            make.centerX.equalToSuperview()
+            make.top.equalTo(stackView.snp_bottomMargin).offset(50)
+            make.left.right.equalTo(stackView)
+        }
+
+        signInButton.snp.makeConstraints { make in
+            make.centerX.equalToSuperview()
+            make.height.equalTo(45)
+            make.width.equalTo(250)
+            make.top.equalTo(errorsLabel).offset(100)
+        }
+    }
+}
+
+extension SignInView: UITextFieldDelegate {
+    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
+        switch textField {
+        case emailTextField:
+            passwordTextField.becomeFirstResponder()
+        case passwordTextField:
+            self.endEditing(true)
+        default:
+            break
+        }
+        return true
+    }
+}
