+++ LoveCalendar/LoveCalendar/Flows/SignIn/Screen/SignInViewController.swift
@@ -11,65 +11,9 @@ import Combine
 
 class SignInViewController: UIViewController, FlowController {
     var completionHandler: (() -> Void)?
+    private let signInView = SignInView(frame: .zero)
     private let viewModel: SignInViewModel
     var cancellables = Set<AnyCancellable>()
-
-    private lazy var signInLabel: UILabel = {
-        let label = UILabel()
-        label.text = "Вход"
-        label.font = .systemFont(ofSize: 20, weight: .bold)
-        label.textColor = UIColor(named: "LabelTextColor")
-        return label
-    }()
-
-    private lazy var emailTextField: UITextField = {
-        let textField = UITextField()
-        textField.placeholder = "Почта"
-        textField.borderStyle = .roundedRect
-        textField.keyboardType = .emailAddress
-        textField.returnKeyType = .next
-        textField.delegate = self
-        textField.textColor = UIColor(named: "LabelTextColor")
-        textField.font = .systemFont(ofSize: 18)
-        return textField
-    }()
-
-    private lazy var passwordTextField: UITextField = {
-        let textField = UITextField()
-        textField.placeholder = "Пароль"
-        textField.borderStyle = .roundedRect
-        textField.isSecureTextEntry = true
-        textField.returnKeyType = .next
-        textField.delegate = self
-        textField.textColor = UIColor(named: "LabelTextColor")
-        textField.font = .systemFont(ofSize: 18)
-        return textField
-    }()
-
-    private lazy var errorsLabel: UILabel = {
-        let label = UILabel()
-        label.font = .systemFont(ofSize: 14)
-        label.textColor = .red
-        label.numberOfLines = 0
-        return label
-    }()
-
-    private lazy var signInButton: UIButton = {
-        let button = UIButton(type: .roundedRect)
-        button.layer.cornerRadius = 8
-        button.titleLabel?.font = .systemFont(ofSize: 18)
-        button.setTitle("Войти", for: .normal)
-        button.setTitleColor(.white, for: .normal)
-        button.backgroundColor = UIColor(named: "ButtonTextColor")
-
-        let action = UIAction { [weak self] _ in
-            guard let self else { return }
-            self.viewModel.signInUser(email: emailTextField.text, password: passwordTextField.text)
-        }
-        button.addAction(action, for: .touchUpInside)
-
-        return button
-    }()
 
     init(viewModel: SignInViewModel) {
         self.viewModel = viewModel
@@ -80,9 +24,14 @@ class SignInViewController: UIViewController, FlowController {
         fatalError("init(coder:) has not been implemented")
     }
 
+    override func loadView() {
+        view = signInView
+    }
+
     override func viewDidLoad() {
+        super.viewDidLoad()
         view.backgroundColor = .white
-        setLayout()
+        signInView.delegate = self
         setBindings()
     }
 }
@@ -93,7 +42,7 @@ extension SignInViewController {
             .receive(on: DispatchQueue.main)
             .sink { [weak self] validationError in
                 guard let self = self else { return }
-                self.errorsLabel.text = validationError
+                self.signInView.setErrors(validationError)
             }
             .store(in: &cancellables)
 
@@ -112,55 +61,14 @@ extension SignInViewController {
             .receive(on: DispatchQueue.main)
             .sink { [weak self] firebaseError in
                 guard let self = self else { return }
-                self.errorsLabel.text = firebaseError
+                self.signInView.setErrors(firebaseError)
             }
             .store(in: &cancellables)
     }
 }
 
-extension SignInViewController {
-    private func setLayout() {
-        let stackView = UIStackView(arrangedSubviews: [emailTextField, passwordTextField])
-        stackView.axis = .vertical
-        stackView.spacing = 20
-
-        addSubviews(signInLabel, stackView, errorsLabel, signInButton)
-
-        signInLabel.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.top.equalToSuperview().offset(view.frame.height * 0.2)
-        }
-
-        stackView.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.width.equalTo(300)
-            make.top.equalTo(signInLabel).offset(view.frame.height * 0.1)
-        }
-
-        errorsLabel.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.top.equalTo(stackView.snp_bottomMargin).offset(50)
-            make.left.right.equalTo(stackView)
-        }
-
-        signInButton.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.width.equalTo(200)
-            make.top.equalTo(errorsLabel).offset(view.frame.height * 0.1)
-        }
-    }
-}
-
-extension SignInViewController: UITextFieldDelegate {
-    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
-        switch textField {
-        case emailTextField:
-            passwordTextField.becomeFirstResponder()
-        case passwordTextField:
-            view.endEditing(true)
-        default:
-            break
-        }
-        return true
+extension SignInViewController: SignViewDelegate {
+    func didPressSignInButton(_ email: String?, _ password: String?) {
+        self.viewModel.signInUser(email: email, password: password)
     }
 }
